backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
pymongo==4.15.5
dnspython==2.8.0
mongoengine==0.29.1
python-dotenv==1.2.1
reportlab==4.4.6
cloudinary==1.44.1
requests==2.32.5
gunicorn==21.2.0
firebase-admin==7.1.0
pywebpush==2.1.2
razorpay==1.4.2
setuptools>=68.0.0
